pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv_pipeline"
version = "0.1.0"
description = "Local pipeline for processing ArXiv papers with graph and vector representation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "requests>=2.28.0",
    "pymongo>=4.3.0",
    "neo4j>=5.5.0",
    "pyyaml>=6.0.1",  # Ensure YAML 1.2+ for better compatibility
    "kaggle>=1.5.12",  # Official Kaggle API client
    "kagglehub>=0.1.0",  # Kaggle model and dataset hub
    "qdrant-client>=1.1.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.30",
    "tqdm>=4.64.0",
    "pillow>=10.0.0",
    "pymupdf>=1.23.0",
    "motor>=3.3.0",
    "transformers",
    "torch>=2.0.0",                 # Base PyTorch package
    "ollama",                        # For local LLM support with Ollama
    "numpy>=1.24.0",               # For data manipulation in notebooks
    "scipy<1.11.0,>=1.9.0",        # Compatible version for Top2Vec
    "bertopic>=0.15.0",             # For topic modeling
    "top2vec>=1.0.0",              # For Top2Vec topic modeling
    "nltk>=3.8.0",                 # For NLP text processing
    "pypdf",                         # Required for PDF loading with LangChain
    "jupyter>=1.0.0",               # Jupyter Notebook support
    "jupyterlab>=4.0.0",            # JupyterLab interface
    "pandas>=2.0.0",                # For data manipulation in notebooks
    "matplotlib>=3.7.0",            # For visualization in notebooks
    "seaborn>=0.12.0",              # For enhanced visualization (needed by Top2Vec)
    "tabulate>=0.9.0",              # For pretty-printing tables in query scripts
    "ipywidgets>=8.0.0",            # Interactive widgets for notebooks
    "python-dotenv>=1.0.0",         # Environment variable support
    "fastapi>=0.100.0",             # API Framework for backend
    "uvicorn>=0.23.0",              # ASGI server for FastAPI
    # "openai>=1.0.0",               # <-- Commented out, not needed for local-only
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]